update demo projects
