+++ demo-jakarta/src/main/java/com/example/demo/model/Book.java
@@ -1,12 +1,12 @@
 package com.example.demo.model;
 
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.util.List;
-import javax.persistence.*;
 
 /**
  * The persistent class for the books database table.
- * 
+ *
  */
 @Entity
 @Table(name="books")
