+++ demo-jakarta/src/main/java/com/example/demo/model/Order.java
@@ -1,12 +1,22 @@
 package com.example.demo.model;
 
-import java.io.Serializable;
-import javax.persistence.*;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+
+import java.io.Serializable;
 
 /**
  * The persistent class for the orders database table.
- * 
+ *
  */
 @Entity
 @Table(name="orders")
