+++ demo/src/main/java/com/example/demo/controller/BookControllerDTO.java
@@ -2,8 +2,8 @@ package com.example.demo.controller;
 
 import com.example.demo.model.Book;
 import com.example.demo.controller.dto.BookDTO;
-import com.example.demo.service.BookService;
-import com.example.demo.service.mapper.BookMapper;
+import com.example.demo.serviceInterface.BookService;
+import com.example.demo.serviceInterface.mapper.BookMapper;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @RestController
 @RequestMapping("/book-dto/")
