+++ demo/src/main/java/com/example/demo/controller/CrudController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 public interface CrudController<O,P>{
