+++ demo/src/main/java/com/example/demo/controller/UserControllerDTO.java
@@ -2,8 +2,8 @@ package com.example.demo.controller;
 
 import com.example.demo.model.User;
 import com.example.demo.controller.dto.UserDTO;
-import com.example.demo.service.UserService;
-import com.example.demo.service.mapper.UserMapper;
+import com.example.demo.serviceInterface.UserService;
+import com.example.demo.serviceInterface.mapper.UserMapper;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @RestController
 @RequestMapping("/user-dto/")
