+++ demo/src/main/java/com/example/demo/repository/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @Repository
 public interface OrderRepository extends JpaRepository<Order, java.lang.Integer> {
