+++ demo/src/main/java/com/example/demo/serviceInterface/CrudService.java
@@ -1,11 +1,11 @@
-package com.example.demo.service;
+package com.example.demo.serviceInterface;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 public interface CrudService<E, P> {
