+++ demo/src/main/java/com/example/demo/serviceInterface/OrderService.java
@@ -1,9 +1,9 @@
-package com.example.demo.service;
+package com.example.demo.serviceInterface;
 
 import com.example.demo.model.Order;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 public interface OrderService extends CrudService<Order,java.lang.Integer> {
