+++ demo/src/main/java/com/example/demo/serviceInterface/UserService.java
@@ -1,9 +1,9 @@
-package com.example.demo.service;
+package com.example.demo.serviceInterface;
 
 import com.example.demo.model.User;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 public interface UserService extends CrudService<User,java.lang.Integer> {
