+++ demo/src/main/java/com/example/demo/serviceInterface/impl/BookServiceBean.java
@@ -1,8 +1,8 @@
-package com.example.demo.service.impl;
+package com.example.demo.serviceInterface.impl;
 
-import com.example.demo.model.Order;
-import com.example.demo.repository.OrderRepository;
-import com.example.demo.service.OrderService;
+import com.example.demo.model.Book;
+import com.example.demo.repository.BookRepository;
+import com.example.demo.serviceInterface.BookService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -12,31 +12,31 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @Service
-public class OrderServiceBean implements OrderService {
+public class BookServiceBean implements BookService {
 
     @Autowired
-    private OrderRepository repository;
+    private BookRepository repository;
 
     @Override
-    public Order create(Order entity) {
+    public Book create(Book entity) {
         return repository.save(entity);
     }
 
     @Override
-    public Order update(Order entity) {
+    public Book update(Book entity) {
         return repository.save(entity);
     }
 
     @Override
-    public Page<Order> read(Pageable pageable) {
+    public Page<Book> read(Pageable pageable) {
         return repository.findAll(pageable);
     }
 
     @Override
-    public Optional<Order> readOne(java.lang.Integer primaryKey) {
+    public Optional<Book> readOne(java.lang.Integer primaryKey) {
         return repository.getOne(primaryKey);
     }
