+++ demo/src/main/java/com/example/demo/serviceInterface/impl/OrderServiceBean.java
@@ -1,8 +1,8 @@
-package com.example.demo.service.impl;
+package com.example.demo.serviceInterface.impl;
 
-import com.example.demo.model.Book;
-import com.example.demo.repository.BookRepository;
-import com.example.demo.service.BookService;
+import com.example.demo.model.Order;
+import com.example.demo.repository.OrderRepository;
+import com.example.demo.serviceInterface.OrderService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -12,31 +12,31 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @Service
-public class BookServiceBean implements BookService {
+public class OrderServiceBean implements OrderService {
 
     @Autowired
-    private BookRepository repository;
+    private OrderRepository repository;
 
     @Override
-    public Book create(Book entity) {
+    public Order create(Order entity) {
         return repository.save(entity);
     }
 
     @Override
-    public Book update(Book entity) {
+    public Order update(Order entity) {
         return repository.save(entity);
     }
 
     @Override
-    public Page<Book> read(Pageable pageable) {
+    public Page<Order> read(Pageable pageable) {
         return repository.findAll(pageable);
     }
 
     @Override
-    public Optional<Book> readOne(java.lang.Integer primaryKey) {
+    public Optional<Order> readOne(java.lang.Integer primaryKey) {
         return repository.getOne(primaryKey);
     }
