+++ demo/src/main/java/com/example/demo/serviceInterface/impl/UserServiceBean.java
@@ -1,8 +1,8 @@
-package com.example.demo.service.impl;
+package com.example.demo.serviceInterface.impl;
 
 import com.example.demo.model.User;
 import com.example.demo.repository.UserRepository;
-import com.example.demo.service.UserService;
+import com.example.demo.serviceInterface.UserService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @Service
 public class UserServiceBean implements UserService {
