+++ demo/src/main/java/com/example/demo/serviceInterface/mapper/UserMapper.java
@@ -1,11 +1,11 @@
-package com.example.demo.service.mapper;
+package com.example.demo.serviceInterface.mapper;
 
 import com.example.demo.model.User;
 import com.example.demo.controller.dto.UserDTO;
 import org.mapstruct.Mapper;
 
 /**
-* Generated by Springboot-3layer-Generator at Feb 20, 2023, 3:17:23 PM
+* Generated by Springboot-3layer-Generator at Feb 21, 2023, 6:23:13 AM
 */
 @Mapper(componentModel = "spring")
 public interface UserMapper {
