+++ src/test/resources/3layer-settings-jakarta.yml
@@ -4,6 +4,7 @@ outputDirectory : src/main/java
 options:
   dtoLayer: true
   serviceInterface: true
+  entityControllers: true
 
 inputPackages:
   jpaEntities: com.example.demo.model
